Moving DeletedUser into Users module space
