+++ spec/models/user_spec.rb
@@ -36,25 +36,6 @@ RSpec.describe User, type: :model do
     allow(Settings::Authentication).to receive(:providers).and_return(Authentication::Providers.available)
   end
 
-  describe "DeletedUser" do
-    subject(:deleted_user) { described_class::DeletedUser }
-
-    it { is_expected.to respond_to(:id) }
-    it { is_expected.to respond_to(:deleted?) }
-    it { is_expected.to respond_to(:darker_color) }
-    it { is_expected.to respond_to(:username) }
-    it { is_expected.to respond_to(:name) }
-    it { is_expected.to respond_to(:summary) }
-    it { is_expected.to respond_to(:twitter_username) }
-    it { is_expected.to respond_to(:github_username) }
-    it { is_expected.to respond_to(:profile_image_url) }
-    it { is_expected.to respond_to(:decorate) }
-    it { is_expected.to respond_to(:path) }
-    it { is_expected.to respond_to(:tag_line) }
-    it { is_expected.to respond_to(:enriched_colors) }
-    it { is_expected.to respond_to(:profile_image_url_for) }
-  end
-
   describe "delegations" do
     it { is_expected.to delegate_method(:admin?).to(:authorizer) }
     it { is_expected.to delegate_method(:any_admin?).to(:authorizer) }
